Code refactor into packages
