+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         </activity>
 
         <activity
-            android:name=".GameActivity"
+            android:name=".game.GameActivity"
             android:exported="false"
             android:label="Game Screen"
             android:theme="@style/Theme.Dicing">
