+++ app/src/main/java/com/sameerasw/dicing/game/components/Dice.kt
@@ -1,4 +1,4 @@
-package com.sameerasw.dicing
+package com.sameerasw.dicing.game.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -20,6 +20,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.sameerasw.dicing.R
 
 @Composable
 fun Dice(
