+++ app/src/main/java/com/sameerasw/dicing/game/components/ScoreHeader.kt
@@ -1,4 +1,4 @@
-package com.sameerasw.dicing.components
+package com.sameerasw.dicing.game.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
